Fix: Parse Discord user IDs as u64 to prevent overflow errors
Discord snowflake IDs are u64 values that can exceed i64::MAX.
Parsing directly to i64 would fail for legitimate high-numbered user IDs.

Solution: Parse as u64 first, then cast to i64 for database storage.
This allows all valid Discord IDs to be processed. Very large IDs
(> i64::MAX) will wrap to negative values but remain unique.

This is a safe short-term fix. For long-term, consider migrating
the database schema to use BIGINT UNSIGNED or storing as String.

Fixes critical issue from Copilot review on PR #8.

+++ backend/src/routes/auth.rs
@@ -66,13 +66,16 @@ pub async fn exchange_code(
         })?;
 
     // Step 3: Parse Discord user ID
+    // Discord IDs are u64 snowflakes, but we store as i64 in the database
+    // Parse as u64 first to handle all valid Discord IDs, then cast to i64
+    // Note: Very large IDs (> i64::MAX) will wrap to negative, but remain unique
     let user_id = discord_user
         .id
-        .parse::<i64>()
+        .parse::<u64>()
         .map_err(|e| {
             tracing::error!("Failed to parse Discord user ID: {}", e);
             StatusCode::INTERNAL_SERVER_ERROR
-        })?;
+        })? as i64;
 
     // Step 4: Create or update user in database
     let avatar_url = discord_user.avatar.as_ref().map(|avatar_hash| {
